Code Performs as Desired
The code now creates an area around the outer edge of the sieve by randomly placing points in the outer holes. The area inside the polygon formed by connecting these outer holes is then stored in an array and the ensemble of areas is plotted in a histogram. Running enough events yields a beautiful Gaussian. Code is generalizable for different sieve designs but only if ncols=nrows. I will try to fix this if I have time. test_poly.C was an attempt to make the TGeoPolygon function draw on a canvas. I failed to make it work so I implemented a TPolyLine that draws the same area as the TGeoPolygon creates. This functionally gives the same visual.